ocr-helper/requirements.txt
# ./ocr-helper/requirements.txt
fastapi
uvicorn
pytesseract
opencv-python-headless
pillow
